Fetch project names from D.o API instead of Update Status API
Some projects randomly disappeared and re-appeared from results, some of them were missing for a long time. See: https://www.drupal.org/project/infrastructure/issues/3373813

On of the downside of this switch is a huge performance drop because instead of switching and processing one huge XML file, we have to make hundreds of API calls.

Also, now that all builds talks to both D.o API and UpdateStatus API, it does not make sense trying to run some of them in parallel.